favorite empty was added
